+++ src/shellyDevice.mock.test.ts
@@ -1057,7 +1057,7 @@ describe('Shelly devices test', () => {
     });
 
     test('Create a gen 2 shellywalldisplay device', async () => {
-      id = 'shellywalldisplay-000822B39A55';
+      id = 'shellywalldisplay-00082261E102';
       log.logName = id;
 
       device = await ShellyDevice.create(shelly, log, path.join('src', 'mock', id + '.json'));
@@ -1065,13 +1065,13 @@ describe('Shelly devices test', () => {
       if (!device) return;
       expect(device.host).toBe(path.join('src', 'mock', id + '.json'));
       expect(device.model).toBe('SAWD-0A1XX10EU1');
-      expect(device.mac).toBe('000822B39A55');
+      expect(device.mac).toBe('00082261E102');
       expect(device.id).toBe(id);
       expect(device.firmware).toBe('2.2.1-eb899af6');
       expect(device.auth).toBe(false);
       expect(device.gen).toBe(2);
       expect(device.profile).toBe(undefined);
-      expect(device.name).toBe('shellywalldisplay-000822B39A55');
+      expect(device.name).toBe('Wall Display');
       expect(device.hasUpdate).toBe(false);
       expect(device.lastseen).not.toBe(0);
       expect(device.online).toBe(true);
@@ -1104,6 +1104,82 @@ describe('Shelly devices test', () => {
       expect(device.getComponent('temperature:0')?.hasProperty('tC')).toBe(true);
       expect(device.getComponent('humidity:0')?.hasProperty('rh')).toBe(true);
       expect(device.getComponent('illuminance:0')?.hasProperty('lux')).toBe(true);
+
+      expect(await device.fetchUpdate()).not.toBeNull();
+
+      if (device) device.destroy();
+    });
+
+    test('Create a gen 2 shellywalldisplay device with thermostat', async () => {
+      id = 'shellywalldisplay-00082261E102';
+      log.logName = id;
+
+      device = await ShellyDevice.create(shelly, log, path.join('src', 'mock', id + '.thermostat.json'));
+      expect(device).not.toBeUndefined();
+      if (!device) return;
+      expect(device.host).toBe(path.join('src', 'mock', id + '.thermostat.json'));
+      expect(device.model).toBe('SAWD-0A1XX10EU1');
+      expect(device.mac).toBe('00082261E102');
+      expect(device.id).toBe(id);
+      expect(device.firmware).toBe('2.2.1-eb899af6');
+      expect(device.auth).toBe(false);
+      expect(device.gen).toBe(2);
+      expect(device.profile).toBe(undefined);
+      expect(device.name).toBe('Wall Display');
+      expect(device.hasUpdate).toBe(false);
+      expect(device.lastseen).not.toBe(0);
+      expect(device.online).toBe(true);
+      expect(device.cached).toBe(false);
+      expect(device.sleepMode).toBe(false);
+
+      expect(device.components.length).toBe(13);
+      expect(device.getComponentNames()).toStrictEqual([
+        'Ble',
+        'WiFi',
+        'Switch',
+        'Input',
+        'Temperature',
+        'Humidity',
+        'Illuminance',
+        'Sys',
+        'Sntp',
+        'Cloud',
+        'MQTT',
+        'WS',
+        'Thermostat',
+      ]);
+      expect(device.getComponentIds()).toStrictEqual([
+        'ble',
+        'wifi_sta',
+        'switch:0',
+        'input:0',
+        'temperature:0',
+        'humidity:0',
+        'illuminance:0',
+        'sys',
+        'sntp',
+        'cloud',
+        'mqtt',
+        'ws',
+        'thermostat:0',
+      ]);
+
+      expect(device.getComponent('sys')?.getValue('temperature')).toBe(undefined);
+      expect(device.getComponent('sys')?.getValue('overtemperature')).toBe(undefined);
+
+      expect(device.getComponent('input:0')?.hasProperty('state')).toBe(true);
+      expect(device.getComponent('switch:0')?.hasProperty('state')).toBe(true);
+
+      expect(device.getComponent('temperature:0')?.hasProperty('tC')).toBe(true);
+      expect(device.getComponent('humidity:0')?.hasProperty('rh')).toBe(true);
+      expect(device.getComponent('illuminance:0')?.hasProperty('lux')).toBe(true);
+      expect(device.getComponent('thermostat:0')?.hasProperty('enable')).toBe(true);
+      expect(device.getComponent('thermostat:0')?.hasProperty('type')).toBe(true);
+      expect(device.getComponent('thermostat:0')?.hasProperty('target_C')).toBe(true);
+      expect(device.getComponent('thermostat:0')?.hasProperty('current_C')).toBe(true);
+      expect(device.getComponent('thermostat:0')?.getValue('target_C')).toBe(25);
+      expect(device.getComponent('thermostat:0')?.getValue('current_C')).toBe(22.9);
+      expect(device.getComponent('thermostat:0')?.getValue('type')).toBe('heating');
 
       expect(await device.fetchUpdate()).not.toBeNull();